pyproject.toml
[tool.poetry]
name = "lichess dwh pipeline"
version = "0.1.0"
description = ""
authors = ["Eden <untouchable.cloud@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
apache-airflow-providers-google = { version = "10.15.0", python = ">=3.10,<3.12" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"